Unit tests need fixing --- WIP
